+++ build-tools/packages/build-cli/package.json
@@ -97,6 +97,7 @@
 		"semver-utils": "^1.1.4",
 		"simple-git": "^3.14.1",
 		"sort-json": "^2.0.1",
+		"sort-package-json": "1.54.0",
 		"strip-ansi": "^6.0.0",
 		"table": "^6.8.0",
 		"type-fest": "^2.19.0"
