--- build-tools/packages/build-tools/bin/fluid-type-test-configure
@@ -1,2 +1,0 @@
-#!/usr/bin/env node
-require("../dist/type-test-configure/main.js");
