+++ build-tools/packages/build-tools/package.json
@@ -21,7 +21,6 @@
 		"fluid-layer-check": "bin/fluid-layer-check",
 		"fluid-repo-policy-check": "bin/fluid-repo-policy-check",
 		"fluid-run-bundle-analyses": "bin/fluid-run-bundle-analyses",
-		"fluid-type-test-configure": "bin/fluid-type-test-configure",
 		"fluid-type-test-generator": "bin/fluid-type-test-generator"
 	},
 	"scripts": {
