+++ build-tools/packages/build-tools/src/index.ts
@@ -34,9 +34,3 @@ export { Handler } from "./repoPolicyCheck/common";
 export { policyHandlers } from "./repoPolicyCheck/handlers";
 export { generateMonoRepoInstallPackageJson } from "./genMonoRepoPackageJson/lib";
 export { type PreviousVersionStyle } from "./common/fluidRepo";
-export {
-	TypeTestConfigActions,
-	updateTypeTestConfiguration,
-	updatePackageJsonFile,
-	VersionOptions,
-} from "./type-test-configure/typeTestConfigUtil";
